+++ cdc/kv/client.go
@@ -39,6 +39,7 @@ import (
 	"github.com/pingcap/ticdc/pkg/version"
 	tidbkv "github.com/pingcap/tidb/kv"
 	"github.com/pingcap/tidb/store/tikv"
+	"github.com/pingcap/tidb/store/tikv/oracle"
 	"github.com/prometheus/client_golang/prometheus"
 	pd "github.com/tikv/pd/client"
 	"go.uber.org/zap"
@@ -114,10 +115,11 @@ type regionStatefulEvent struct {
 }
 
 type regionFeedState struct {
-	sri                   singleRegionInfo
-	requestID             uint64
-	regionEventCh         chan *regionEvent
-	stopped               int32
+	sri           singleRegionInfo
+	requestID     uint64
+	regionEventCh chan *regionEvent
+	stopped       int32
+
 	initialized           bool
 	matcher               *matcher
 	startFeedTime         time.Time
@@ -781,9 +783,9 @@ MainLoop:
 				limiter := s.client.getRegionLimiter(regionID)
 				g.Go(func() error {
 					if !s.enableRegionWorker {
-						return s.singleReceiveFromStream(ctx, g, rpcCtx.Addr, getStoreID(rpcCtx), stream, pendingRegions, limiter)
+						return s.receiveFromStream(ctx, g, rpcCtx.Addr, getStoreID(rpcCtx), stream, pendingRegions, limiter)
 					}
-					return s.receiveFromStream(ctx, g, rpcCtx.Addr, getStoreID(rpcCtx), stream, pendingRegions, limiter)
+					return s.receiveFromStreamV2(ctx, g, rpcCtx.Addr, getStoreID(rpcCtx), stream, pendingRegions, limiter)
 				})
 			}
 
@@ -851,6 +853,85 @@ func needReloadRegion(failStoreIDs map[uint64]struct{}, rpcCtx *tikv.RPCContext)
 		}
 	}
 	return
+}
+
+// partialRegionFeed establishes a EventFeed to the region specified by regionInfo.
+// It manages lifecycle events of the region in order to maintain the EventFeed
+// connection. If any error happens (region split, leader change, etc), the region
+// and error info will be sent to `errCh`, and the receiver of `errCh` is
+// responsible for handling the error and re-establish the connection to the region.
+func (s *eventFeedSession) partialRegionFeed(
+	ctx context.Context,
+	state *regionFeedState,
+	limiter *rate.Limiter,
+) error {
+	receiver := state.regionEventCh
+	defer func() {
+		state.markStopped()
+		// Workaround to avoid remaining messages in the channel blocks the receiver thread.
+		// TODO: Find a better solution.
+		timer := time.After(time.Second * 2)
+		for {
+			select {
+			case <-receiver:
+			case <-timer:
+				return
+			}
+		}
+	}()
+
+	ts := state.sri.ts
+	maxTs, err := s.singleEventFeed(ctx, state.sri.verID.GetID(), state.sri.span, state.sri.ts, receiver)
+	log.Debug("singleEventFeed quit")
+
+	if err == nil || errors.Cause(err) == context.Canceled {
+		return nil
+	}
+
+	failpoint.Inject("kvClientErrUnreachable", func() {
+		if err == errUnreachable {
+			failpoint.Return(err)
+		}
+	})
+
+	if maxTs > ts {
+		ts = maxTs
+	}
+
+	regionID := state.sri.verID.GetID()
+	log.Info("EventFeed disconnected",
+		zap.Uint64("regionID", regionID),
+		zap.Uint64("requestID", state.requestID),
+		zap.Stringer("span", state.sri.span),
+		zap.Uint64("checkpoint", ts),
+		zap.String("error", err.Error()))
+
+	state.sri.ts = ts
+
+	// We need to ensure when the error is handled, `isStopped` must be set. So set it before sending the error.
+	state.markStopped()
+
+	failpoint.Inject("kvClientSingleFeedProcessDelay", nil)
+
+	now := time.Now()
+	delay := limiter.ReserveN(now, 1).Delay()
+	if delay != 0 {
+		log.Info("EventFeed retry rate limited",
+			zap.Duration("delay", delay), zap.Reflect("regionID", regionID))
+		t := time.NewTimer(delay)
+		defer t.Stop()
+		select {
+		case <-t.C:
+			// We can proceed.
+		case <-ctx.Done():
+			return ctx.Err()
+		}
+	}
+
+	return s.onRegionFail(ctx, regionErrorInfo{
+		singleRegionInfo: state.sri,
+		err:              err,
+	})
 }
 
 // divideAndSendEventFeedToRegions split up the input span into spans aligned
@@ -998,11 +1079,11 @@ func (s *eventFeedSession) receiveFromStream(
 ) error {
 	// Cancel the pending regions if the stream failed. Otherwise it will remain unhandled in the pendingRegions list
 	// however not registered in the new reconnected stream.
-	var wg sync.WaitGroup
 	defer func() {
 		log.Info("stream to store closed", zap.String("addr", addr), zap.Uint64("storeID", storeID))
 
 		remainingRegions := pendingRegions.takeAll()
+
 		for _, state := range remainingRegions {
 			err := s.onRegionFail(ctx, regionErrorInfo{
 				singleRegionInfo: state.sri,
@@ -1013,42 +1094,14 @@ func (s *eventFeedSession) receiveFromStream(
 				return
 			}
 		}
-
-		wg.Wait()
-
-		s.workersLock.Lock()
-		delete(s.workers, addr)
-		s.workersLock.Unlock()
 	}()
 
 	captureAddr := util.CaptureAddrFromCtx(ctx)
 	changefeedID := util.ChangefeedIDFromCtx(ctx)
 	metricSendEventBatchResolvedSize := batchResolvedEventSize.WithLabelValues(captureAddr, changefeedID)
 
-	s.workersLock.Lock()
-	worker, ok := s.workers[addr]
-	if !ok {
-		worker = &regionWorker{
-			session:        s,
-			limiter:        limiter,
-			inputCh:        make(chan *regionStatefulEvent, 1024),
-			outputCh:       s.eventCh,
-			regionStates:   make(map[uint64]*regionFeedState),
-			enableOldValue: s.enableOldValue,
-		}
-		s.workers[addr] = worker
-	}
-	s.workersLock.Unlock()
-
-	wg.Add(1)
-	go func() {
-		err := worker.run(ctx)
-		if err != nil && errors.Cause(err) != context.Canceled && !cerror.ErrEventFeedAborted.Equal(err) {
-			// TODO: should this error be propagated?
-			log.Error("run region worker failed", zap.Error(err))
-		}
-		wg.Done()
-	}()
+	// Each region has it's own goroutine to handle its messages. `regionStates` stores states of these regions.
+	regionStates := make(map[uint64]*regionFeedState)
 
 	for {
 		cevent, err := stream.Recv()
@@ -1056,8 +1109,11 @@ func (s *eventFeedSession) receiveFromStream(
 		failpoint.Inject("kvClientStreamRecvError", func() {
 			err = errors.New("injected stream recv error")
 		})
+		// TODO: Should we have better way to handle the errors?
 		if err == io.EOF {
-			close(worker.inputCh)
+			for _, state := range regionStates {
+				close(state.regionEventCh)
+			}
 			return nil
 		}
 		if err != nil {
@@ -1087,11 +1143,12 @@ func (s *eventFeedSession) receiveFromStream(
 			// with SendMsg, in future refactor we should refine the recv loop
 			s.deleteStream(addr)
 
-			// send nil regionStatefulEvent to signal worker exit
-			select {
-			case worker.inputCh <- nil:
-			case <-ctx.Done():
-				return ctx.Err()
+			for _, state := range regionStates {
+				select {
+				case state.regionEventCh <- nil:
+				case <-ctx.Done():
+					return ctx.Err()
+				}
 			}
 
 			// Do no return error but gracefully stop the goroutine here. Then the whole job will not be canceled and
@@ -1111,14 +1168,14 @@ func (s *eventFeedSession) receiveFromStream(
 		}
 
 		for _, event := range cevent.Events {
-			err = s.sendRegionChangeEvent(ctx, g, event, worker, pendingRegions, addr, limiter)
+			err = s.sendRegionChangeEvent(ctx, g, event, regionStates, pendingRegions, addr, limiter)
 			if err != nil {
 				return err
 			}
 		}
 		if cevent.ResolvedTs != nil {
 			metricSendEventBatchResolvedSize.Observe(float64(len(cevent.ResolvedTs.Regions)))
-			err = s.sendResolvedTs(ctx, g, cevent.ResolvedTs, worker, addr)
+			err = s.SendResolvedTs(ctx, g, cevent.ResolvedTs, regionStates, pendingRegions, addr)
 			if err != nil {
 				return err
 			}
@@ -1130,12 +1187,12 @@ func (s *eventFeedSession) sendRegionChangeEvent(
 	ctx context.Context,
 	g *errgroup.Group,
 	event *cdcpb.Event,
-	worker *regionWorker,
+	regionStates map[uint64]*regionFeedState,
 	pendingRegions *syncRegionFeedStateMap,
 	addr string,
 	limiter *rate.Limiter,
 ) error {
-	state, ok := worker.regionStates[event.RegionId]
+	state, ok := regionStates[event.RegionId]
 	// Every region's range is locked before sending requests and unlocked after exiting, and the requestID
 	// is allocated while holding the range lock. Therefore the requestID is always incrementing. If a region
 	// is receiving messages with different requestID, only the messages with the larges requestID is valid.
@@ -1172,25 +1229,12 @@ func (s *eventFeedSession) sendRegionChangeEvent(
 			return cerror.ErrNoPendingRegion.GenWithStackByArgs(event.RegionId, event.RequestId, addr)
 		}
 
-		state.start()
 		// Then spawn the goroutine to process messages of this region.
-		worker.regionStates[event.RegionId] = state
+		regionStates[event.RegionId] = state
 
-		// send resolved event when starting a single event feed
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		case s.eventCh <- &model.RegionFeedEvent{
-			RegionID: state.sri.verID.GetID(),
-			Resolved: &model.ResolvedSpan{
-				Span:       state.sri.span,
-				ResolvedTs: state.sri.ts,
-			},
-		}:
-		}
-		// g.Go(func() error {
-		// 	return s.partialRegionFeed(ctx, state, limiter)
-		// })
+		g.Go(func() error {
+			return s.partialRegionFeed(ctx, state, limiter)
+		})
 	} else if state.isStopped() {
 		log.Warn("drop event due to region feed stopped",
 			zap.Uint64("regionID", event.RegionId),
@@ -1200,25 +1244,25 @@ func (s *eventFeedSession) sendRegionChangeEvent(
 	}
 
 	select {
+	case state.regionEventCh <- &regionEvent{
+		changeEvent: event,
+	}:
 	case <-ctx.Done():
 		return ctx.Err()
-	case worker.inputCh <- &regionStatefulEvent{
-		changeEvent: event,
-		state:       state,
-	}:
 	}
 	return nil
 }
 
-func (s *eventFeedSession) sendResolvedTs(
+func (s *eventFeedSession) SendResolvedTs(
 	ctx context.Context,
 	g *errgroup.Group,
 	resolvedTs *cdcpb.ResolvedTs,
-	worker *regionWorker,
+	regionStates map[uint64]*regionFeedState,
+	pendingRegions *syncRegionFeedStateMap,
 	addr string,
 ) error {
 	for _, regionID := range resolvedTs.Regions {
-		state, ok := worker.regionStates[regionID]
+		state, ok := regionStates[regionID]
 		if ok {
 			if state.isStopped() {
 				log.Warn("drop resolved ts due to region feed stopped",
@@ -1228,9 +1272,8 @@ func (s *eventFeedSession) sendResolvedTs(
 				continue
 			}
 			select {
-			case worker.inputCh <- &regionStatefulEvent{
+			case state.regionEventCh <- &regionEvent{
 				resolvedTs: resolvedTs,
-				state:      state,
 			}:
 			case <-ctx.Done():
 				return ctx.Err()
@@ -1238,6 +1281,241 @@ func (s *eventFeedSession) sendResolvedTs(
 		}
 	}
 	return nil
+}
+
+// singleEventFeed handles events of a single EventFeed stream.
+// Results will be send to eventCh
+// EventFeed RPC will not return resolved event directly
+// Resolved event is generate while there's not non-match pre-write
+// Return the maximum resolved
+func (s *eventFeedSession) singleEventFeed(
+	ctx context.Context,
+	regionID uint64,
+	span regionspan.ComparableSpan,
+	startTs uint64,
+	receiverCh <-chan *regionEvent,
+) (uint64, error) {
+	captureAddr := util.CaptureAddrFromCtx(ctx)
+	changefeedID := util.ChangefeedIDFromCtx(ctx)
+	metricEventSize := eventSize.WithLabelValues(captureAddr)
+	metricPullEventInitializedCounter := pullEventCounter.WithLabelValues(cdcpb.Event_INITIALIZED.String(), captureAddr, changefeedID)
+	metricPullEventCommittedCounter := pullEventCounter.WithLabelValues(cdcpb.Event_COMMITTED.String(), captureAddr, changefeedID)
+	metricPullEventCommitCounter := pullEventCounter.WithLabelValues(cdcpb.Event_COMMIT.String(), captureAddr, changefeedID)
+	metricPullEventPrewriteCounter := pullEventCounter.WithLabelValues(cdcpb.Event_PREWRITE.String(), captureAddr, changefeedID)
+	metricPullEventRollbackCounter := pullEventCounter.WithLabelValues(cdcpb.Event_ROLLBACK.String(), captureAddr, changefeedID)
+	metricSendEventResolvedCounter := sendEventCounter.WithLabelValues("native-resolved", captureAddr, changefeedID)
+	metricSendEventCommitCounter := sendEventCounter.WithLabelValues("commit", captureAddr, changefeedID)
+	metricSendEventCommittedCounter := sendEventCounter.WithLabelValues("committed", captureAddr, changefeedID)
+
+	initialized := false
+
+	matcher := newMatcher()
+	advanceCheckTicker := time.NewTicker(time.Second * 5)
+	defer advanceCheckTicker.Stop()
+	lastReceivedEventTime := time.Now()
+	startFeedTime := time.Now()
+	lastResolvedTs := startTs
+	handleResolvedTs := func(resolvedTs uint64) error {
+		if !initialized {
+			return nil
+		}
+		if resolvedTs < lastResolvedTs {
+			log.Warn("The resolvedTs is fallen back in kvclient",
+				zap.String("Event Type", "RESOLVED"),
+				zap.Uint64("resolvedTs", resolvedTs),
+				zap.Uint64("lastResolvedTs", lastResolvedTs),
+				zap.Uint64("regionID", regionID))
+			return nil
+		}
+		// emit a checkpointTs
+		revent := &model.RegionFeedEvent{
+			RegionID: regionID,
+			Resolved: &model.ResolvedSpan{
+				Span:       span,
+				ResolvedTs: resolvedTs,
+			},
+		}
+		lastResolvedTs = resolvedTs
+
+		select {
+		case s.eventCh <- revent:
+			metricSendEventResolvedCounter.Inc()
+		case <-ctx.Done():
+			return errors.Trace(ctx.Err())
+		}
+		return nil
+	}
+
+	select {
+	case s.eventCh <- &model.RegionFeedEvent{
+		RegionID: regionID,
+		Resolved: &model.ResolvedSpan{
+			Span:       span,
+			ResolvedTs: startTs,
+		},
+	}:
+	case <-ctx.Done():
+		return lastResolvedTs, errors.Trace(ctx.Err())
+	}
+	resolveLockInterval := 20 * time.Second
+	failpoint.Inject("kvClientResolveLockInterval", func(val failpoint.Value) {
+		resolveLockInterval = time.Duration(val.(int)) * time.Second
+	})
+
+	for {
+		var event *regionEvent
+		var ok bool
+		select {
+		case <-ctx.Done():
+			return lastResolvedTs, ctx.Err()
+		case <-advanceCheckTicker.C:
+			if time.Since(startFeedTime) < resolveLockInterval {
+				continue
+			}
+			if !s.isPullerInit.IsInitialized() {
+				// Initializing a puller may take a long time, skip resolved lock to save unnecessary overhead.
+				continue
+			}
+			sinceLastEvent := time.Since(lastReceivedEventTime)
+			if sinceLastEvent > resolveLockInterval {
+				log.Warn("region not receiving event from tikv for too long time",
+					zap.Uint64("regionID", regionID), zap.Stringer("span", span), zap.Duration("duration", sinceLastEvent))
+			}
+			version, err := s.kvStorage.(*StorageWithCurVersionCache).GetCachedCurrentVersion()
+			if err != nil {
+				log.Warn("failed to get current version from PD", zap.Error(err))
+				continue
+			}
+			currentTimeFromPD := oracle.GetTimeFromTS(version.Ver)
+			sinceLastResolvedTs := currentTimeFromPD.Sub(oracle.GetTimeFromTS(lastResolvedTs))
+			if sinceLastResolvedTs > resolveLockInterval && initialized {
+				log.Warn("region not receiving resolved event from tikv or resolved ts is not pushing for too long time, try to resolve lock",
+					zap.Uint64("regionID", regionID), zap.Stringer("span", span),
+					zap.Duration("duration", sinceLastResolvedTs),
+					zap.Uint64("resolvedTs", lastResolvedTs))
+				maxVersion := oracle.ComposeTS(oracle.GetPhysical(currentTimeFromPD.Add(-10*time.Second)), 0)
+				err = s.lockResolver.Resolve(ctx, regionID, maxVersion)
+				if err != nil {
+					log.Warn("failed to resolve lock", zap.Uint64("regionID", regionID), zap.Error(err))
+					continue
+				}
+			}
+			continue
+		case event, ok = <-receiverCh:
+		}
+		if !ok {
+			log.Debug("singleEventFeed receiver closed")
+			return lastResolvedTs, nil
+		}
+
+		if event == nil {
+			log.Debug("singleEventFeed closed by error")
+			return lastResolvedTs, cerror.ErrEventFeedAborted.GenWithStackByArgs()
+		}
+		lastReceivedEventTime = time.Now()
+		if event.changeEvent != nil {
+			metricEventSize.Observe(float64(event.changeEvent.Event.Size()))
+			switch x := event.changeEvent.Event.(type) {
+			case *cdcpb.Event_Entries_:
+				for _, entry := range x.Entries.GetEntries() {
+					switch entry.Type {
+					case cdcpb.Event_INITIALIZED:
+						if time.Since(startFeedTime) > 20*time.Second {
+							log.Warn("The time cost of initializing is too mush",
+								zap.Duration("timeCost", time.Since(startFeedTime)),
+								zap.Uint64("regionID", regionID))
+						}
+						metricPullEventInitializedCounter.Inc()
+						initialized = true
+						cachedEvents := matcher.matchCachedRow()
+						for _, cachedEvent := range cachedEvents {
+							revent, err := assembleRowEvent(regionID, cachedEvent, s.enableOldValue)
+							if err != nil {
+								return lastResolvedTs, errors.Trace(err)
+							}
+							select {
+							case s.eventCh <- revent:
+								metricSendEventCommitCounter.Inc()
+							case <-ctx.Done():
+								return lastResolvedTs, errors.Trace(ctx.Err())
+							}
+						}
+					case cdcpb.Event_COMMITTED:
+						metricPullEventCommittedCounter.Inc()
+						revent, err := assembleRowEvent(regionID, entry, s.enableOldValue)
+						if err != nil {
+							return lastResolvedTs, errors.Trace(err)
+						}
+
+						if entry.CommitTs <= lastResolvedTs {
+							logPanic("The CommitTs must be greater than the resolvedTs",
+								zap.String("Event Type", "COMMITTED"),
+								zap.Uint64("CommitTs", entry.CommitTs),
+								zap.Uint64("resolvedTs", lastResolvedTs),
+								zap.Uint64("regionID", regionID))
+							return lastResolvedTs, errUnreachable
+						}
+						select {
+						case s.eventCh <- revent:
+							metricSendEventCommittedCounter.Inc()
+						case <-ctx.Done():
+							return lastResolvedTs, errors.Trace(ctx.Err())
+						}
+					case cdcpb.Event_PREWRITE:
+						metricPullEventPrewriteCounter.Inc()
+						matcher.putPrewriteRow(entry)
+					case cdcpb.Event_COMMIT:
+						metricPullEventCommitCounter.Inc()
+						if entry.CommitTs <= lastResolvedTs {
+							logPanic("The CommitTs must be greater than the resolvedTs",
+								zap.String("Event Type", "COMMIT"),
+								zap.Uint64("CommitTs", entry.CommitTs),
+								zap.Uint64("resolvedTs", lastResolvedTs),
+								zap.Uint64("regionID", regionID))
+							return lastResolvedTs, errUnreachable
+						}
+						ok := matcher.matchRow(entry)
+						if !ok {
+							if !initialized {
+								matcher.cacheCommitRow(entry)
+								continue
+							}
+							return lastResolvedTs, cerror.ErrPrewriteNotMatch.GenWithStackByArgs(entry.GetKey(), entry.GetStartTs())
+						}
+
+						revent, err := assembleRowEvent(regionID, entry, s.enableOldValue)
+						if err != nil {
+							return lastResolvedTs, errors.Trace(err)
+						}
+
+						select {
+						case s.eventCh <- revent:
+							metricSendEventCommitCounter.Inc()
+						case <-ctx.Done():
+							return lastResolvedTs, errors.Trace(ctx.Err())
+						}
+					case cdcpb.Event_ROLLBACK:
+						metricPullEventRollbackCounter.Inc()
+						matcher.rollbackRow(entry)
+					}
+				}
+			case *cdcpb.Event_Admin_:
+				log.Info("receive admin event", zap.Stringer("event", event.changeEvent))
+			case *cdcpb.Event_Error:
+				return lastResolvedTs, cerror.WrapError(cerror.ErrEventFeedEventError, &eventError{err: x.Error})
+			case *cdcpb.Event_ResolvedTs:
+				if err := handleResolvedTs(x.ResolvedTs); err != nil {
+					return lastResolvedTs, errors.Trace(err)
+				}
+			}
+		}
+
+		if event.resolvedTs != nil {
+			if err := handleResolvedTs(event.resolvedTs.Ts); err != nil {
+				return lastResolvedTs, errors.Trace(err)
+			}
+		}
+	}
 }
 
 func (s *eventFeedSession) addStream(storeAddr string, stream cdcpb.ChangeData_EventFeedClient) {
